+++ src/constants/platform/url.ts
@@ -211,6 +211,9 @@ export const CannotConnectAccountUrl =
 export const LockedConnectAccountUrl =
   'https://storage.googleapis.com/geyser-projects-media/utils/connect-account-locked.png'
 
+export const VerifyEmailImageUrl =
+  'https://storage.googleapis.com/geyser-projects-media/utils/enter-otp.svg'
+
 // GRANTS
 
 export const Grant3AnnouncementImageUrl =