+++ src/pages/otp/VerifyYourEmailContent.tsx
@@ -1,0 +1,115 @@
+import { Button, Image } from '@chakra-ui/react'
+import { useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import { BsArrowLeft } from 'react-icons/bs'
+
+import { Body1 } from '../../components/typography'
+import { VerifyEmailImageUrl } from '../../constants'
+import {
+  MfaAction,
+  OtpResponseFragment,
+  useSendOtpByEmailMutation,
+} from '../../types'
+import { useNotification } from '../../utils'
+import { ReceiveOneTimePassword, VerifyOneTimePassword } from './components'
+
+interface VerifyYourEmailContentProps {
+  action: MfaAction
+  handleVerify?: (
+    otpCode: number,
+    optData: OtpResponseFragment,
+    email?: string,
+  ) => void
+}
+
+export const VerifyYourEmailContent = ({
+  action,
+  handleVerify,
+}: VerifyYourEmailContentProps) => {
+  const { t } = useTranslation()
+  const { toast } = useNotification()
+
+  const [sentOtp, setSentOtp] = useState(false)
+  const [otpData, setOtpData] = useState<OtpResponseFragment>()
+  const [inputEmail, setInputEmail] = useState('')
+
+  const [sendOtpByEmail] = useSendOtpByEmailMutation({
+    onError() {
+      toast({
+        status: 'error',
+        title: 'Failed to generate otp.',
+        description: 'Please try again',
+      })
+    },
+    onCompleted(data) {
+      const otp = data.sendOTPByEmail
+      if (otp) {
+        setSentOtp(true)
+        setOtpData(otp)
+      }
+    },
+  })
+
+  const handleSendOtpByEmail = (email: string) => {
+    sendOtpByEmail({
+      variables: {
+        input: {
+          action,
+          email,
+        },
+      },
+    })
+  }
+
+  const getDescription = () => {
+    if (action === MfaAction.ProjectWalletUpdate) {
+      return t(
+        'You can update your wallet securely by using the One Time Password sent to your verified email.',
+      )
+    }
+
+    return t(
+      'Backup your Geyser account and project with your email. This will ensure that you can always access Geyser (in case of social media censorship) and can securely update your project information.',
+    )
+  }
+
+  return (
+    <>
+      {sentOtp && (
+        <Button
+          variant="secondary"
+          size={{ base: 'sm', lg: 'md' }}
+          leftIcon={<BsArrowLeft fontSize="20px" />}
+          onClick={() => setSentOtp(false)}
+          alignSelf="start"
+        >
+          {'Back'}
+        </Button>
+      )}
+      <Image
+        src={VerifyEmailImageUrl}
+        alt="verify-email-image"
+        w={200}
+        h={200}
+        alignSelf="center"
+      />
+
+      <Body1 semiBold>{getDescription()}</Body1>
+      {sentOtp && otpData ? (
+        <VerifyOneTimePassword
+          action={action}
+          otp={otpData}
+          handleSendOtpByEmail={handleSendOtpByEmail}
+          inputEmail={inputEmail}
+          handleVerify={handleVerify}
+        />
+      ) : (
+        <ReceiveOneTimePassword
+          action={action}
+          setInputEmail={setInputEmail}
+          handleSendOtpByEmail={handleSendOtpByEmail}
+        />
+      )}
+    </>
+  )
+}